chore: rename statistic-test to statistical-analysis for api folder
