+++ src/modules/statistical-analysis/components/statistic-test/results.tsx
@@ -1,7 +1,7 @@
 import { Stack } from '@mantine/core';
 import React from 'react';
 import { BaseStatisticTestResultRendererProps } from '../../types';
-import { StatisticTestResultModel } from '@/api/statistic-test';
+import { StatisticTestResultModel } from '@/api/statistical-analysis';
 import {
   GroupCountsRenderer,
   SignificanceAndEffectSizeComponents,
