+++ src/modules/statistical-analysis/data-provider/statistic-test.tsx
@@ -1,7 +1,7 @@
 import {
   PairwiseStatisticTestResultModel,
   StatisticTestResultModel,
-} from '@/api/statistic-test';
+} from '@/api/statistical-analysis';
 import { BaseStatisticTestDataProviderHook } from '../types';
 import { client } from '@/common/api/client';
 import {
